Start splitting NVI menu into its own component
