+++ src/pages/messages/components/NviMenuContent.tsx
@@ -1,0 +1,261 @@
+import AdjustIcon from '@mui/icons-material/Adjust';
+import {
+  Box,
+  FormControlLabel,
+  FormLabel,
+  LinearProgress,
+  Link as MuiLink,
+  Radio,
+  Typography,
+  styled,
+} from '@mui/material';
+import { useState } from 'react';
+import { useTranslation } from 'react-i18next';
+import { Link, useHistory } from 'react-router-dom';
+import { useFetchNviCandidates } from '../../../api/hooks/useFetchNviCandidates';
+import { FetchNviCandidatesParams, NviCandidatesSearchParam } from '../../../api/searchApi';
+import { NavigationListAccordion } from '../../../components/NavigationListAccordion';
+import { StyledTicketSearchFormGroup } from '../../../components/styled/Wrappers';
+import { TasksPageLocationState } from '../../../types/locationState.types';
+import { NviCandidateSearchStatus } from '../../../types/nvi.types';
+import { ROWS_PER_PAGE_OPTIONS } from '../../../utils/constants';
+import { dataTestId } from '../../../utils/dataTestIds';
+import { UrlPathTemplate } from '../../../utils/urlPaths';
+
+const StyledStatusRadio = styled(Radio)({
+  paddingTop: '0.05rem',
+  paddingBottom: '0.05rem',
+});
+
+const StyledNviStatusBox = styled(Box)({
+  padding: '0.5rem',
+  borderRadius: '0.25rem',
+});
+
+export const NviMenuContent = () => {
+  const { t } = useTranslation();
+  const history = useHistory<TasksPageLocationState>();
+  const searchParams = new URLSearchParams(history.location.search);
+
+  const isOnNviCandidatesPage = history.location.pathname === UrlPathTemplate.TasksNvi;
+  const isOnNviStatusPage = history.location.pathname === UrlPathTemplate.TasksNviStatus;
+
+  const queryParam = searchParams.get(NviCandidatesSearchParam.Query);
+  const assigneeParam = searchParams.get(NviCandidatesSearchParam.Assignee);
+  const nviStatusParam = searchParams.get(NviCandidatesSearchParam.Filter) as NviCandidateSearchStatus | null;
+  const nviStatusFilter = nviStatusParam ?? 'pending';
+  const nviYearParam = searchParams.get(NviCandidatesSearchParam.Year);
+  const nviYearFilter = nviYearParam ? +nviYearParam : new Date().getFullYear();
+
+  const setNviStatusParam = (newStatusFilter: NviCandidateSearchStatus) => {
+    searchParams.set(NviCandidatesSearchParam.Filter, newStatusFilter);
+    history.push({ search: searchParams.toString() });
+  };
+
+  const openCandidatesView = () => {
+    if (!isOnNviCandidatesPage) {
+      history.push(UrlPathTemplate.TasksNvi);
+    }
+  };
+
+  const affiliationsParam = searchParams.get(NviCandidatesSearchParam.Affiliations)?.split(',');
+
+  const commonNviParams = {
+    affiliations: affiliationsParam,
+    assignee: assigneeParam,
+    query: queryParam,
+    year: nviYearFilter,
+  } satisfies FetchNviCandidatesParams;
+
+  const nviAggregationsQuery = useFetchNviCandidates({
+    enabled: isOnNviCandidatesPage || isOnNviStatusPage,
+    params: { ...commonNviParams, size: 1, aggregation: 'all' },
+  });
+
+  const [page, setPage] = useState(1);
+  const [rowsPerPage, setRowsPerPage] = useState(ROWS_PER_PAGE_OPTIONS[0]);
+
+  const sizeParam = searchParams.get(NviCandidatesSearchParam.Size);
+  const sizeValue = sizeParam ? +sizeParam : rowsPerPage;
+
+  const listNviCandidatesParams = {
+    ...commonNviParams,
+    size: sizeValue,
+    offset: (page - 1) * sizeValue,
+    filter: nviStatusFilter,
+  } satisfies FetchNviCandidatesParams;
+
+  const nviCandidatesQuery = useFetchNviCandidates({
+    params: listNviCandidatesParams,
+  });
+
+  const nviAggregations = nviAggregationsQuery.data?.aggregations;
+
+  const nviPendingCount = nviAggregations?.pending.docCount.toLocaleString();
+  const nviPendingCollaborationCount = nviAggregations?.pendingCollaboration.docCount.toLocaleString();
+  const nviAssignedCount = nviAggregations?.assigned.docCount.toLocaleString();
+  const nviAssignedCollaborationCount = nviAggregations?.assignedCollaboration.docCount.toLocaleString();
+  const nviApprovedCount = nviAggregations?.approved.docCount.toLocaleString();
+  const nviApprovedCollaborationCount = nviAggregations?.approvedCollaboration.docCount.toLocaleString();
+  const nviRejectedCount = nviAggregations?.rejected.docCount.toLocaleString();
+  const nviRejectedCollaborationCount = nviAggregations?.rejectedCollaboration.docCount.toLocaleString();
+
+  const nviCandidatesTotal = nviAggregations?.totalCount.docCount ?? 0;
+  const nviCandidatesCompeted = nviAggregations?.completed.docCount ?? 0;
+  const nviCompletedPercentage =
+    nviCandidatesTotal > 0 ? Math.round((nviCandidatesCompeted / nviCandidatesTotal) * 100) : 100;
+
+  return (
+    <NavigationListAccordion
+      title={t('common.nvi')}
+      startIcon={<AdjustIcon sx={{ bgcolor: 'nvi.main' }} />}
+      accordionPath={UrlPathTemplate.TasksNvi}
+      onClick={() => {
+        if (!isOnNviCandidatesPage) {
+          setPage(1);
+        }
+      }}
+      dataTestId={dataTestId.tasksPage.nviAccordion}>
+      <StyledTicketSearchFormGroup>
+        {nviAggregationsQuery.isSuccess && (
+          <StyledNviStatusBox sx={{ bgcolor: 'nvi.light', mb: '1rem' }}>
+            <Typography id="progress-label" gutterBottom>
+              {t('tasks.nvi.completed_count', {
+                completed: nviCandidatesCompeted,
+                total: nviCandidatesTotal,
+              })}
+            </Typography>
+            <LinearProgress
+              aria-labelledby="progress-label"
+              variant="determinate"
+              value={nviCompletedPercentage}
+              sx={{
+                my: '0.175rem',
+                height: '0.75rem',
+                bgcolor: 'white',
+              }}
+            />
+            <Typography sx={{ textAlign: 'center' }}>{nviCompletedPercentage} %</Typography>
+
+            <Box sx={{ display: 'flex', justifyContent: 'center', mt: '0.5rem' }}>
+              <MuiLink
+                data-testid={dataTestId.tasksPage.nvi.toggleStatusLink}
+                component={Link}
+                to={isOnNviCandidatesPage ? UrlPathTemplate.TasksNviStatus : UrlPathTemplate.TasksNvi}>
+                {isOnNviCandidatesPage ? t('tasks.nvi.show_reporting_status') : t('tasks.nvi.show_candidate_search')}
+              </MuiLink>
+            </Box>
+          </StyledNviStatusBox>
+        )}
+
+        <FormLabel component="legend" sx={{ fontWeight: 700 }}>
+          {t('tasks.status')}
+        </FormLabel>
+
+        <Box sx={{ display: 'flex', flexDirection: 'column', gap: '0.25rem' }}>
+          <StyledNviStatusBox sx={{ bgcolor: 'nvi.light' }}>
+            <FormControlLabel
+              data-testid={dataTestId.tasksPage.nvi.statusFilter.pendingRadio}
+              checked={nviStatusFilter === 'pending'}
+              disabled={!!assigneeParam}
+              onClick={openCandidatesView}
+              control={<StyledStatusRadio onChange={() => setNviStatusParam('pending')} />}
+              slotProps={{ typography: { fontWeight: 700 } }}
+              label={nviPendingCount ? `${t('tasks.nvi.status.New')} (${nviPendingCount})` : t('tasks.nvi.status.New')}
+            />
+            <FormControlLabel
+              data-testid={dataTestId.tasksPage.nvi.statusFilter.pendingCollaborationRadio}
+              checked={nviStatusFilter === 'pendingCollaboration'}
+              disabled={!!assigneeParam}
+              onClick={openCandidatesView}
+              control={<StyledStatusRadio onChange={() => setNviStatusParam('pendingCollaboration')} />}
+              label={
+                nviPendingCount
+                  ? `${t('tasks.nvi.waiting_for_your_institution')} (${nviPendingCollaborationCount})`
+                  : t('tasks.nvi.waiting_for_your_institution')
+              }
+            />
+          </StyledNviStatusBox>
+
+          <StyledNviStatusBox sx={{ bgcolor: 'nvi.light' }}>
+            <FormControlLabel
+              data-testid={dataTestId.tasksPage.nvi.statusFilter.assignedRadio}
+              checked={nviStatusFilter === 'assigned'}
+              onClick={openCandidatesView}
+              control={<StyledStatusRadio onChange={() => setNviStatusParam('assigned')} />}
+              slotProps={{ typography: { fontWeight: 700 } }}
+              label={
+                nviAssignedCount
+                  ? `${t('tasks.nvi.status.Pending')} (${nviAssignedCount})`
+                  : t('tasks.nvi.status.Pending')
+              }
+            />
+            <FormControlLabel
+              data-testid={dataTestId.tasksPage.nvi.statusFilter.assignedCollaborationRadio}
+              checked={nviStatusFilter === 'assignedCollaboration'}
+              onClick={openCandidatesView}
+              control={<StyledStatusRadio onChange={() => setNviStatusParam('assignedCollaboration')} />}
+              label={
+                nviAssignedCollaborationCount
+                  ? `${t('tasks.nvi.waiting_for_your_institution')} (${nviAssignedCollaborationCount})`
+                  : t('tasks.nvi.waiting_for_your_institution')
+              }
+            />
+          </StyledNviStatusBox>
+
+          <StyledNviStatusBox sx={{ bgcolor: 'secondary.main' }}>
+            <FormControlLabel
+              data-testid={dataTestId.tasksPage.nvi.statusFilter.approvedRadio}
+              checked={nviStatusFilter === 'approved'}
+              onClick={openCandidatesView}
+              control={<StyledStatusRadio onChange={() => setNviStatusParam('approved')} />}
+              slotProps={{ typography: { fontWeight: 700 } }}
+              label={
+                nviApprovedCount
+                  ? `${t('tasks.nvi.status.Approved')} (${nviApprovedCount})`
+                  : t('tasks.nvi.status.Approved')
+              }
+            />
+            <FormControlLabel
+              data-testid={dataTestId.tasksPage.nvi.statusFilter.approvedCollaborationRadio}
+              checked={nviStatusFilter === 'approvedCollaboration'}
+              onClick={openCandidatesView}
+              control={<StyledStatusRadio onChange={() => setNviStatusParam('approvedCollaboration')} />}
+              label={
+                nviApprovedCollaborationCount
+                  ? `${t('tasks.nvi.waiting_for_other_institutions')} (${nviApprovedCollaborationCount})`
+                  : t('tasks.nvi.waiting_for_other_institutions')
+              }
+            />
+          </StyledNviStatusBox>
+
+          <StyledNviStatusBox sx={{ bgcolor: 'secondary.main' }}>
+            <FormControlLabel
+              data-testid={dataTestId.tasksPage.nvi.statusFilter.rejectedRadio}
+              checked={nviStatusFilter === 'rejected'}
+              onClick={openCandidatesView}
+              control={<StyledStatusRadio onChange={() => setNviStatusParam('rejected')} />}
+              slotProps={{ typography: { fontWeight: 700 } }}
+              label={
+                nviRejectedCount
+                  ? `${t('tasks.nvi.status.Rejected')} (${nviRejectedCount})`
+                  : t('tasks.nvi.status.Rejected')
+              }
+            />
+            <FormControlLabel
+              data-testid={dataTestId.tasksPage.nvi.statusFilter.rejectedCollaborationRadio}
+              checked={nviStatusFilter === 'rejectedCollaboration'}
+              onClick={openCandidatesView}
+              control={<StyledStatusRadio onChange={() => setNviStatusParam('rejectedCollaboration')} />}
+              label={
+                nviRejectedCollaborationCount
+                  ? `${t('tasks.nvi.waiting_for_other_institutions')} (${nviRejectedCollaborationCount})`
+                  : t('tasks.nvi.waiting_for_other_institutions')
+              }
+            />
+          </StyledNviStatusBox>
+        </Box>
+      </StyledTicketSearchFormGroup>
+    </NavigationListAccordion>
+  );
+};
